api/services/database_seeding_service/requirements.txt
dnspython
mysql-connector-python
pandas
tldextract
numpy
requests
beautifulsoup4
nodriver
mysqlx
python-dotenv
